FindMaxima method seems to be working.
Threw a few pathological test cases at FindMaxima, and it seemed to work. Needs more rigorous testing. Ultimately, needed two 'visited' arrays in the method. The first is to prevent the outer loop from searching stuff we've already searched in a lake and decided the fate of. The second is used exclusively in lakes; the way the algorithm is implemented, it is possible to explore only part of a lake, and stop. In that case, we need to be able to search previously visited parts of lakes in order to determine whether the current lake is a max or not. So the outer 'visited' array won't suffice. Therefore the second array keeps track of the pixels we've searched in the current lake.