Made findMaxima for 2D and 3D faster using my way (queues) and not with the array.
Using an idea from trying to implement Michael Schmid's technique, I now prevent the same object from being added to the toSearch LL multiple times with an additional check. Sped it up to as fast as Michael's implementation I did. Not sure how it compares to the existing findMaxima exactly, but it's very fast...