Added downsampling to the findMaxima algorithm to improve run time.
Used the DownSample class from Imglib to reduce the numer of pixels in the image. Given a blob diameter from the user, the image is downsized in order to get the blobs to be about 10 pixels in diameter, at which point a non-rigorously tested "optimal" sigma of 1.55 is used (gave the best results from trial-and-error) in the gaussian filter applied thereafter.