Embryo_Tracker: tracking now deals with starting/ending new track segment.
Previously, it was possible to ignore track segments of size 1, which is bad in case it actually belongs to a longer track. Now, if a solution is found in the top right quadrant, it is labeled as a segment of size one (and overwritten to the normal case if later found in the top left corner). If a solution is in the bottom left corner, we just ignore it, because the default value is NOT_LINKED, which signals the end of a segment.

Signed-off-by: Nick Perry <nicholas.m.perry@gmail.com>