TrackMate_: added a GUI panel to make an initial threshold
In some case we might get such a high number of spots that it would be
unwise to try to calculate their features or display them.
Now a panel appearing just after the segmentation step allow to discard
spots whose quality (a feature set by the segmenter at creation) is too
low (or too high).
Then all features are calculated, and the graphical way to threshold
spots is displayed.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>