TrackMate_: moved to using a weigthed graph
Tracks are now represented by a weighted graph, which will allow to
store the cost of each link (edge) - we do not want to discard it since
it costed us so much to compute it. Also, it might be useful in the
future to make graph operation on the tracking result.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>