TrackMate_: added GUI panel to choose segmenter, tracker, displayer
The user can now choose what he want just before it is used. For
instance, the choice of trackers is only proposed to him before actually
doing the tracking.
Before that, onw had to choose at the beginning, meaning that to change
the tracker, one had to go back to the start, and redo the segmentation.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>