TrackScheme_: fix for NPE in trackscheme when creating new spots
There was a NPE thrown when redoing the layout in the track-scheme,
after a spot has been created in the hyperstack displayer.

This was due because the view in the track-scheme is not in sync with
the the model graph. When a spot was added in another class, the
track-scheme had its model updated, but this was not reflected in the
view, because of the layer of JGraph atop of the JGraphT.

The fix was to register the track-sheme as a listener for change in the
model occurring in the displayer, and generating view when needed.

While at it, the zoom in track scheme now almost works correctly.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>