TrackMate_: more work on the track model
The track model is actually more complicated than I thought.
Tracks must be generated on the fly from the underlying graph, since it
is the graph that register new link creations or deletions. Since we
want to to modify it in a GUI at random times, the use cases became
quite complicated.
To try to deal with that, I used the update model that can be found in
the JGraphX project, where an update level is increased every time a
modification is made to the model. When the modification is done, it is
decreased by 1. So multiple modifications  will result in a model fields
re-calculation only when this update level reaches 0, which saves some
time.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>