TrackMate_: consolidate model
The model must be able to deal with situation where the user does a
manual edit in the graph. This manual edit can trigger changes in the
graph that were not done by the user:

By complex change, we mean the changes occurring in the graph caused by
another change that was initiated manually by the user. For instance, imagine
we have a simple track branch made of 5 spots that link linearly, like this:

 S1 - S2 - S3 - S4 - S5

The user might want to remove the S3 spot, in the middle of the track. On top
of the track rearrangement, that is dealt with elsewhere in the model
class, this spot removal also triggers 2 edges removal: the links S2-S3 and S3-S4
disappear. The only way for the {@link TrackMateModel} to be aware of that,
and to forward these events to its listener, is to listen itself to the
{@link #graph} that store links.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>