TrackMate_: fix a bug in tracking: hanging ig all costs are blocked
If the top-left corner of the linking cost matrix was all set to
blocking value - which happens when the user set a distance cutoff
smaller than the minimal distance between two spots accross two frames -
the hungarian solver was hanging.
This commit fix this bug by handling this special case in the LAP
tracker class.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>