TrackMate_: fix a severe bug in tracking, causing feature costs to be ignored.
Normally, when the user configures the settings for the trackers, he can
choose to add features weight in the linking cost calculation. This is
aimed as privileging the links between spots that resemble each other
the most.
However, this was forgotten, and then ignored, in the linking cost
calculation. This commit fixes this, using a way to calculate the
penalty that is nice, works well, and will be propagated to other cost
calculation (gap closing, merging, splitting).
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>