TrackMate_: a solution to save settings object to XML
We work by adding 2 methods in the 2 settings interfaces (one for
segmenter settings and one for tracking settings).
- void marshall(Element) saves the concrete settings to the JDom element
- and void unmarshall(Element) does the opposite and set the fields of
  the concrete object from the JDom element.

Concrete instance are guessed from a common attribute in the main
element.

This means that each concrete settings object has to specify how to
marshall and unmarshall itself. So on top of mixing GUI elements with
settings object, we now add JDom elements... well...
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>