TrackMate_: update JGraphX to 1.8.0
There is a bug I cannot fix with TrackScheme:
When I use TrackMate from Fiji, built with the Fiji build system, it
seems that the cells in the TrackScheme cannot be moved.
However, if I run it from Eclipse, they are movable.
I do not undertand this, particularly because I am on a Mac, and the JVM
used are the same, right?
Updating JGraphX did not solve this...

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>