TrackMate_: fix a nasty bug in saving class names of tracker settings
We now rely on .getClass().getName() to save (amongst others) the name
of the tracker settings we used.
The only problem is that it does not work well with anonymous classes.
And I was using such anonymous classes for the simple tracker, changing
a method on the fly to return the simplified config panel.
To wor around this, I now have a public method in the settings class
that allow to specofy if we want to return the simplfied panel or the
normal one. The class name is then saved normally.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>