TrackMate_: fix a bug where track features were null after loading
We save the track features in the file, but we do not retrieve them when
loading. Instead, we recalculate them with the current track analyzers.
The main problem was that track analyzers are declared only in the class
TrackMate_, so calling to analyze tracks from the newly loaded model had
no effect (no track analyzers are there).
Fix this by calling to analyze tracks from the newly loaded PLUGIN.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>