TrackMate_: use a more clever way for retrieving tracker settings
If managed to retrieve and instantiate the tracker, we can get a default
settings from it. We can then update this settings content with the
file, assuming the file matches the default settings class.
If not, we priviliege what is saved in the file.
I should actually do that for segmenter settings as well....
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>