TrackMate_: move plugin init code to GUI part, shorten track analyzers list
The TrackMate_ class is the entry point for the plugin, whether you use
it in scripts or in the GUI. However, we only use the lists of modules
in the GUI case. Initiating these lists take a very long time. So we
will instantiate them only when we use the GUI.

Also, remove 2 track feature analyzers from these lists, for they take
ages to compute.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>