TrackMate_: fix the track visibility after loading bug
To fix this bug, I had to add two methods to TmXmlReader and 2 methods
to TrackMateModel to read and override the tracks as list of spots and
slit of edges.

It is a bit dirty, because you have now 2 methods in the model that can
override 2 fields that are normally generated on the fly from the graph.
Also, the 2 extra methods in the reader parse the file at the same place
that for the graph...

But this is all I have for now.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>