TrackMate_: much better memory management in LAP tracker
The second step in Jaqama et al. LAP tracker requires to build a matrix
for costs associated with gap-closing, merging and splitting events.

This yields a gigantic matrix, as soon as the total number of spots is
big. However, when we forbid merging and splitting events, we prune this
matrix a lot after constructing it.

This is nonsense: the cost matrix can so big that it generates an
OutOfMemory error, even if it is going to be shrunk to 1% when pruning.

To avoid this, we simple create a smaller matrix right away, and skip
merging and splitting cost calculation if we can.

So when we have to deal only with gap-closing events, not only we save a
lot of memory, but we also go much faster by skipping these useless
steps.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>