TrackMate_: fix a bug in the GUI when filtering spots
The filtering GUI for spots had some features values that were empty
even if they existed in the spots. It turns out it was because for this
GUI we gather the spot feature values in a multi-threaded way, but were
using a non-concurrent hash map. Using a concurrent hash map fixed the
problem.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>