TrackMate_: fix a convoluted bug in TrackScheme with missing edges
The bug was manifesting when launching TrackScheme after creating lonely
spots (that were therefore not shown in TrackScheme), then launching
TrackScheme, then creating links targetting these invisible spots.
The edge to the firstly invisible spots, made visible by linking, was
invisible (the edge, not the spot). That is because it had a geometry
targetting a wrong cells.

The fix consisted in ensuring that the source and target of a link
existed in TrackScheme prior to importing a new edge to it.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>