TrackMate_: remove generic contamination
I was using a type T everywhere: the type propagated to classes
that did not need to be generic, just because of a few declarations.
I decided to sacrifice cleanliness and have a few raw types and
unchecked warnings rather than nasty imglib2 dependence everywhere.

This is all good and well, but now my code does not compile with the
build system, though it does in Eclipse. :/
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>