TrackMate_: better monitoring of removed edges.
We can always find to what track and edge has been added, and therefore
recompute features for this track only.

But what happens if we remove an edge? If this generates two new tracks,
we need to know their IDs. But the new IDs are going to be different
than the old ID the removed edge belonged to.

This commit adds a system to retrieve the new target track IDs that are
affected by an edge removal. The TrackMateModelChangeEvent can give this
information for removed edges (only).
The flushUpdate and recomputeTrackFromGraph in the TrackMateModel and
TrackGraphModel classes respectively generate this information, in an
unfortunately rather convoluted way.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>