TrackMate_: change the track feature update mechanism
Less intelligence: each track feature analyzer needs to declare whether
it is local or not. Local means that if one track changes, we just need
to recompute the feature for this track. Non-local means that for any
change to the tracks, the feature needs to be re-calculated over all the
filtered tracks.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>