TrackMate_: fix bug preventing track merging and splitting in some cases
Noticed by Jan Brocher:
We were using the POSITION_T to check if 2 spots were in the next or
previous frame when determining if they are elligible to a splitting or
merging event.
This would of course fail as soon as the time calibration put more than 1
in units of time between frames. For instance a frame interval of 5
seconds would make any splitting or merging events not-detected.

The proper way to do it is to use the FRAME feature of each spot, which
correctly reports the frame number, whatever is the physical frame
interval.

While at it, use the FRAME feature anywhere we can.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>