TrackMate_: do not use directed graphs
It turns out this was a bad idea.
It is error prone because we might create by mistake two links between
the same two spots, one in each direction. We then also would have to
test for the existence of 2 links every time etc...
It is much mre safer to add specialized iterators and classes that can
generate a kind of -on-the-fly directed graph, using the Spot#FRAME
feature to get a iterator directed in time. A directed graph just repeat
this information more clumsily in our case.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>