TrackMate_: start developing v2.2.0
This minor version will focus on API changes to facility scripting or
re-using TrackMate as library.
API changes normally foster a major version number change. However,
these changes are expected to be transparent to the end user (TrackMate
is mainly used through the GUI), so we do not want to introduce false
hopes there.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>