TrackMate_: fix longstanding bug with 3D viewer
When using the 3D viewer as the main view in theGUI, the console was
often prompted for weird (to me) java3D error, such as:

javax.media.j3d.NnuIdManager replace SEVERE: Can't Find matching nnuId.

I thought it was due to the spot or track content being displayed, but
actually the problem was caused by the way we added the image content to
the 3D universe: I was using universe.addVoltex(), whereas adding
content should be done in an axync manner.
Using ContentCreater.createConten() then universe.addContentLater()
fixed the problem.
Now there are other issues of course.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>