TrackMate_: fix a serious mistake in LoG.
The laplacian filter was not intialized properly: because the kernel was
calculated using integers - whereas floats were expected - it was
equivalent to a single white point, hereby doing nothing.
Therefore, our LoG filter was actually just a Gaussian filter.
That changes a lot for the actual results.
This is terrible, people have been using this for years and I only
noticed now. Arrrhhh.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>