TrackMate_: change the way the ROI is used internally.
Prior to this commit, we were relyingon building an ImgPlus that was
cropped from the source ImgPlus. Using a specialized class, one could
avoid copygin the data.

Now we rely on passing the full image (still, just one timepoint), and
an interval that determines the ROI. I had to change the
SpotDetectorFactory interface to do that,

The benefit is that it fixes a severe bug in the DogDetector when using
a ROI (reported by Philippe Mailly & JÃrÃmie Teillon). Also, we now use
more of the "native" imglib2 classes and interface rather than custom
classes.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>