TrackMate_: fix several serious problems with the LoG kernel.
1.
In the case where we have a strong anisotropic calibration (the pixel
sizes are very different e.g. in X and Z), we CANNOT generate a faithful
LoG kernel by simply scaling the sigmas in each direction and working in
pixel coordinates. If we do so, the LoG kernel does not look like a
mexican hat anymore.

For instance, if the pixel sizes are { 0.2, 0.2, 1 } um/pixel (as in my
C.elegans data by the way), then the LoG kernel built with the old
fashion resembles a dull Gaussian in the central XY plane, and to a
sinusoid in the central XZ plane.
I found out this to be very detrimental to the accuracy of detection:
several spurious spots could be obtained for a single image blob. Tiny
ones could be missed.

The right way is to do all the kernel calculations using image
coordinates with a single sigma determined from the specified blob
radius, in image coordinates. Then the LoG kernel has the shape of a LoG
kernel, scaled anisotropically in space.

2.
The normalization constant was not right, which in 3D caused the optimal
sigma not to be determined accuratly from the specified blob radius. It
was too small, which effectively resulted in having maximal responses
for blob larger than expected.

3.
I added a fudge factor so that the quality values are close to the
source image values.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>