Simplify Spot constructors.
Using a double[] as a parameter was a *stupid* idea. The double[] array
is often instantiated just to create the spot, and is forgotten just
after that. So we end up creating a lot of useless arrays, for the sake
of nothing.
Whether we could reuse a RealLocalizable, or just use 3 plain double
primitives.
This commit just does that.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>