TrackMate: DWIM swapped T/Z
When time and z are swapped, there is typically only one timepoint
(when there should be many). Detect the situation and ask the user
whether they wanted to swap the dimensions, but also offer them to
leave the dimensions as-are, or to cancel.

While at it, fix a frame interval of 0 seconds (because later parts
of TrackMate get confused by that) to 1sec/frame, as suggested by
Jean-Yves Tinevez.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>