Use SciJava-common's annotation indexer to discover tracker classes
The coolest idea of TrackMate is to make it an extensible platform for
tracking. To make it truly extensible, we need to have a common interface
for all tracker algorithms and a way to discover them so that TrackMate
can be extended even by third-party .jar files without touching TrackMate
itself at all.

Luckily, such a framework exists: SciJava-common provides it in the form
of its PluginService and the annotation indexer.

Basically all one needs to do to make a plugin of a certain type
discoverable at runtime is to annotate the respective class using

	@Plugin(type = ACertainType.class)

All the application that wants to discover thusly annotated classes at
runtime is to make sure that the "certain type" is defined as an interface
that extends the SciJavaPlugin interface.

To discover the classes, a SciJava context's PluginService can be queried
for them:

	Context context = new Context(PluginService.class);
	PluginService pluginService = context.getService(PluginService);
	for (PluginInfo<ACertainType> info :
			pluginService.getPluginsOfType(ACertainType.class)) {
		...
	}

Magic!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>