Feature analyzers can now be sorted by their priority annotation.
The fantastic PluginInfo we are using to discover feature analyzers has
a priority annotation that accepts a double as value. We use this value
to tag the priority of the analyzer.

Some analyzers rely on other analyzer features. So the depending
analyzers must be called AFTER the analyzer they depend on. We use the
priority value to sort them.

An analyzer that does not depend on anything will have - by convention -
a priority value of 0.
An analyzer that depends on other analyzers will have a priority equals
to the largest priority of the analyzers it depends on, plus one.

So far, we only use this for spot feature analyzers.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>