User factories for actioms.
We need to do this, since we need for SciJava to offer a blank
constructor. But some of the TM actions needs to be passed a
TrackMateGUIController, which must be done before we call
execute(TrackMate) on them. So we go through a factory, even small,
which also makes things more elegant.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>