Fine-tune multithreading for detection.
Before this commit, TrackMate was doing multithreading during detection
in the following way:

We run the detection process on as many frames as we have threads, and
each frame is processed using 1 thread.

This is suboptimal if we have a large number of threads, but a few
number of frames (e.g. 1 frame, 24 threads). This commit fixes this:

If we have 10 threads and 15 frames to process, we process 10 frames at
once, and allocate 1 thread per frame. But if we have 10 threads and 2
frames, we process the 2 frames at once, and allocate 5 threads per
frame if we can.

For this to work of course, the detector must be multithreaded, which is
the case with the ones we ship now.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>