WIP: Manual color selection for spots and edges.
The logic to store and assign a color to a selection of edges and spot
is there. It exploits a SpotFeatureAnalyzerFactory and an EdgeAnalyzer
that declare a single feature and attribute it a default double value
when called.

This double value is casted to an int, which is used to get the RGB of a
color.

Some actions in the popup menu of TrackScheme allows to modify manually
this feature through a JColorChooser dialog.

When the manual feature is selected in the ColorByFeatureGuiPanel, some
special actions are taken: the color generators are not per feature per
se, but two new color generator that retrieve the double value stored
for the manual color, cast it to int, and return the matching color.

TODO:
- Have TrackScheme shows the manual spot color when selecting the manual
  edge color (or not, we could live with just the manual edge color
  here)
- Check that the manual color value is saved and restored to./from XML.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>