Fix loading XML file.
The spot IDs were not assigned properly.
They were loaded properly, but the Spot constructor used was assigning
the ID to the frame field, which caused the ID of the newly create spot
to be determined automatically.
When building tracks, the reader could therefore not found the spots
with the IDs written in the XML.
This is fixed by re-engineering a bit the spot constructors.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>