Add track filter "longest gap"
This allows filtering tracks by their longest track. A use case is when we
have unreliable spot detection and want to exclude tracks that are likely
to be too imprecise to be useful for analysis: after closing way too large
gaps, we can exclude those tracks based on their longest gap.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>