Fix NPE when saving with a source image not loaded from disk.
When trying to save to XML, a NPE was thrown in the case the user was
working with an image that was not present on the hardrive (such as
after duplication of another image, or as a result of another plugin
process).
On this image, the getOriginalInfo() method returns null, which caused
the NPE when trying to retrieve a folder to offer to the user.
This commit just adds some safety check and default to the user folder
in case of such an image.

This commit alone does not fixes the bug. It happens somewhere else
after.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>

Conflicts:
	src/main/java/fiji/plugin/trackmate/gui/descriptors/SaveDescriptor.java