Do not crash when saving unconfigured detector or tracker.
If the detectorFactory or the trackerFactory of the settings objects are
null, saving would crash TrackMate.
This is not OK as we want to allow saving before they are defined by the
user, so that they can resume their work at any state in the GUI.
This commit fixes the crash.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>