Disallow default penalties from being modified
When the map containing the default penalties is used directly, that means
that it gets modified when adding penalties. As a consequence, subsequent
runs will inherit settings from whatever previous run.

In particular with the LAPTrackerTest, this wreaked havoc because the
first test added penalties for the mean intensity while the second test
did not even provide the mean intensity, resulting in a nasty
NullPointerException.

Thanks, Jeniece Jenkins.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>