Release v2.3.0.
- TrackScheme has 2 levels of decoration.

- It is possible to manually color spots and edges. The colors
attributed are saved and retrieved from XML.

- There is now an action to export to Icy track manager. It generates a
XML file that can be opened in the Icy track manager.

- Icy XML track files can be opened in TrackMate. Just point the
TrackMate loader plugin to such a file, and its type will be detected
and it will be opened in TrackMate.

- A new action allows merging two TrackMate files into one.

- The color scale used to display colored tracks, links and spots can
now be manually adjusted. In the display panel of the GUI, double click
on the label "color by" and a settings window will open allowing the
user to select between manual or automatic scaling. This is great to
generate views that are comparable between datasets. This feature is
commissioned and offered by Fumio Hayashi, Kobe University.

- Fix a potential (and potentially long standing) bug with numerical
input.

- The table in the InfoPane of TrackScheme displays the full spot name.

- Auto-linking mode allows the creation of links backward in time.

- Two new track feature analyzers allow filtering tracks by their spots
quality or by their longest gap.

- Put back the forgotten Manual Tracker.

- Fix crash triggered when the specified ROI had points out of the source image bounds.

- Fix crash when saving with a source image not loaded from disk.

- Do not crash when saving unconfigured detector or tracker.

- Many other bugfixes.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>