Fix a TrackScheme bug when linking to invisible tracks.
The bug was the following:

Some tracks are marked as not-visible, for instance after filtering
tracks using track filters. These invisible tracks are not displayed in
TrackScheme. Which means that their edges and spots do not have a
matching JGraphX cell.

When the user manually creates a link between a spot and another spot
which belings to an invisible track, the invisible track becomes visible
by convention. All of it.

TrackScheme was confused afterwards, because there were spots and edges
in the model belonging to visible tracks for which it did not have
JGraphX cells.

The fix consists of checking whether the edge the user created link to
an existing cell. If not, we create a cell for each spot and edge of the
track if required.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>