Add an interactive test that show differences between trackers.
Between the sparse and the non-sparse.
The differences originate from the cost threshold: in the non-sparse
version, it is solely based on the distance between two spots. In the
sparse version, it is based on the cost, but we still specify a max
distance. This is for settings compatibility, that sacrifices results
identity.

However they are minor in practice: 99.73% of the edges are found
identifical with both tracker. The tracker accuracy is surely well below
this value.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>