Rename and reorder trackers in the GUI.
The dense formulation (by Nick) of the LAP tracker framework is now
called "Old LAP tracker" and co. The tracker keys are unchanged, so it
is ensured that upon loading a XML file, the right tracker will be
retrieved.

The sparse formulation are now called "LAP tracker" and co, with new
keys, and are the default.

The order is not determined using the "priority" tag of the SciJava
annotation.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>