Do not register the spot edit twice.
A bug, noticed by petersfbn@googlemail.com, prevented to properly use
the manual tracking tools when laucnhing them a second time in the same
Fiji session.

This was caused by the fact that when getting the spot edit tool
instance, we were registering the target image twice the second time it
was launched.

This fixes the bug #956 in bugzilla.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>