Fix a bug in the Linear motion LAP tracker.
The Kalman tracker class could fail on some peculiar cases:
If two Kalma filters were making the exact same prediction in position,
the compareTo methods would deem them as equal (the method implementation
compares the position of the prediction). Having two different objects
that are equal according to compareTo generated a duplicate assignment in
the DefaultCostMatrixCreateor, which crashed it.

This solution consists in returning the difference of hashCodes when the
two positions are found equal, so that two different objects are still
different even if they point to the same position.

Now of course, the question is how could TWO Kalman filters have the same
EXACT prediction. There might be something odd here I could not track.

Noticed by Ronny Sczech, thanks to him.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>